Move IT noncompliant example to ManuallyAddedNoncompliantIssues
